+++ ExamplePlugin/Skills/Death/DeathState.cs
@@ -8,10 +8,12 @@ using EntityStates;
 using R2API;
 using static R2API.DamageAPI;
 using VoidJailerMod.Damage;
+using RoR2;
+using Xan.ROR2VoidPlayerCharacterCommon.EntityStates;
+using Xan.ROR2VoidPlayerCharacterCommon;
 
 namespace VoidJailerMod.Skills.Death {
-	public class DeathState : GenericCharacterDeath {
-		public override bool shouldAutoDestroy => false;
+	public class DeathState : VoidDeathStateBase {
 
 		public override void PlayDeathAnimation(float crossfadeDuration = 0.1f) {
 			PlayCrossfade("Body", "Death", "Death.playbackRate", Duration, crossfadeDuration);
@@ -22,7 +24,7 @@ namespace VoidJailerMod.Skills.Death {
 			muzzleTransform = FindModelChild(MuzzleName);
 			if (muzzleTransform && isAuthority) {
 				FireProjectileInfo fireProjectileInfo = default;
-				fireProjectileInfo.projectilePrefab = ProjectileProvider.ExaggeratedDeathBombProjectile;
+				fireProjectileInfo.projectilePrefab = VoidImplosionObjects.JailerImplosion;
 				fireProjectileInfo.position = muzzleTransform.position;
 				fireProjectileInfo.rotation = Quaternion.LookRotation(characterDirection.forward, Vector3.up);
 				fireProjectileInfo.owner = gameObject;